Added basic messaging-broadcast system
Needs a lot of work as it currently only gets the last message in the
relay’s static buffer. I will need to investigate using core.async due
to the fact that to render the messages require they be constantly
update, and therefore as soon as we reset the static buffer, it’ll
vanish as it just rendered.
Maybe there’s a function i can call to make it not render?