Hydra tweaks
The Hydra is now four tiles long with the addition of hydra-rear. Also,
the hydra boss room has been changed to a desert for easier navigation.
We can turn it back to a maze on a further iteration, I just felt this
to be the quicker solution considering the sprint ends soon.